+++ GPTifier/src/command_embed.cpp
@@ -4,9 +4,9 @@
 #include "cli.hpp"
 #include "configs.hpp"
 #include "datadir.hpp"
-#include "input_selection.hpp"
 #include "params.hpp"
 #include "parsers.hpp"
+#include "selectors.hpp"
 #include "utils.hpp"
 
 #include <fmt/core.h>
